+++ src/models/CommerceClient.ts
@@ -7,10 +7,19 @@ export type APICredentials = {
   accessToken: string;
 };
 
+export type HttpHandlerOptions = {
+  connectionTimeout: number;
+  socketTimeout: number;
+  keepAlive: boolean;
+  family: 4;
+};
+
 export interface CommerceClientConfig {
   credentials?: Partial<AppCredentials> & Partial<APICredentials>;
+  httpHandler?: Partial<HttpHandlerOptions>;
 }
 
 export interface CommerceClientResolvedConfig {
   credentials: AppCredentials & APICredentials;
+  httpHandler: HttpHandlerOptions;
 }